// src/frontend-applications/agent-portal/modules/faq/model/FAQCategoryProperty.ts
/**
 * Copyright (C) 2006-2023 c.a.p.e. IT GmbH, https://www.cape-it.de
 * --
 * This software comes with ABSOLUTELY NO WARRANTY. For details, see
 * the enclosed file LICENSE for license information (GPL3). If you
 * did not receive this file, see https://www.gnu.org/licenses/gpl-3.0.txt.
 * --
 */

export enum FAQCategoryProperty {

    ID = 'ID',

    NAME = 'Name',

    COMMENT = 'Comment',

    PARENT_ID = 'ParentID',

    GROUP_IDS = 'GroupIDs',

    VALID_ID = 'ValidID',

    CREATE_BY = 'CreateBy',

    CREATE_TIME = 'CreateTime',

    CHANGE_BY = 'ChangeBy',

    CHANGE_TIME = 'ChangeTime',

    SUB_CATEGORIES = 'SubCategories',

    FULL_NAME = 'Fullname',

    ICON = 'ICON',

    ARTICLES = 'Articles'

}